// src/app/api/model/movieCredits200Response.ts
/**
 * tmdb-api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MovieCredits200ResponseCastInner } from './movieCredits200ResponseCastInner';
import { MovieCredits200ResponseCrewInner } from './movieCredits200ResponseCrewInner';


export interface MovieCredits200Response { 
    id?: number;
    cast?: Array<MovieCredits200ResponseCastInner>;
    crew?: Array<MovieCredits200ResponseCrewInner>;
}