// src/app/api/model/movieDetails200Response.ts
/**
 * tmdb-api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MovieDetails200ResponseProductionCompaniesInner } from './movieDetails200ResponseProductionCompaniesInner';
import { MovieDetails200ResponseProductionCountriesInner } from './movieDetails200ResponseProductionCountriesInner';
import { MovieDetails200ResponseGenresInner } from './movieDetails200ResponseGenresInner';
import { MovieDetails200ResponseSpokenLanguagesInner } from './movieDetails200ResponseSpokenLanguagesInner';


export interface MovieDetails200Response { 
    adult?: boolean;
    backdrop_path?: string;
    belongs_to_collection?: any | null;
    budget?: number;
    genres?: Array<MovieDetails200ResponseGenresInner>;
    homepage?: string;
    id?: number;
    imdb_id?: string;
    original_language?: string;
    original_title?: string;
    overview?: string;
    popularity?: number;
    poster_path?: string;
    production_companies?: Array<MovieDetails200ResponseProductionCompaniesInner>;
    production_countries?: Array<MovieDetails200ResponseProductionCountriesInner>;
    release_date?: string;
    revenue?: number;
    runtime?: number;
    spoken_languages?: Array<MovieDetails200ResponseSpokenLanguagesInner>;
    status?: string;
    tagline?: string;
    title?: string;
    video?: boolean;
    vote_average?: number;
    vote_count?: number;
}