// src/app/api/model/movieImages200Response.ts
/**
 * tmdb-api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { MovieImages200ResponseBackdropsInner } from './movieImages200ResponseBackdropsInner';
import { MovieImages200ResponseLogosInner } from './movieImages200ResponseLogosInner';
import { MovieImages200ResponsePostersInner } from './movieImages200ResponsePostersInner';


export interface MovieImages200Response { 
    backdrops?: Array<MovieImages200ResponseBackdropsInner>;
    id?: number;
    logos?: Array<MovieImages200ResponseLogosInner>;
    posters?: Array<MovieImages200ResponsePostersInner>;
}