// src/app/api/model/tvSeasonDetails200ResponseEpisodesInner.ts
/**
 * tmdb-api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TvSeasonDetails200ResponseEpisodesInnerGuestStarsInner } from './tvSeasonDetails200ResponseEpisodesInnerGuestStarsInner';
import { TvSeasonDetails200ResponseEpisodesInnerCrewInner } from './tvSeasonDetails200ResponseEpisodesInnerCrewInner';


export interface TvSeasonDetails200ResponseEpisodesInner { 
    air_date?: string;
    episode_number?: number;
    episode_type?: string;
    id?: number;
    name?: string;
    overview?: string;
    production_code?: string;
    runtime?: number;
    season_number?: number;
    show_id?: number;
    still_path?: string;
    vote_average?: number;
    vote_count?: number;
    crew?: Array<TvSeasonDetails200ResponseEpisodesInnerCrewInner>;
    guest_stars?: Array<TvSeasonDetails200ResponseEpisodesInnerGuestStarsInner>;
}